+++ src/Form/Type/NewsletterSubscriptionFormType.php
@@ -24,7 +24,7 @@ final class NewsletterSubscriptionFormType extends AbstractType
                     'required' => true,
                     'attr' => [
                         'placeholder' => 'newsletter.form.name',
-                        'class' => 'lp-c-light-grey bg-transparent border-2 rounded-0',
+                        'class' => 'lp-c-light-grey border-2 rounded-0',
                     ],
                 ]
             )
@@ -36,7 +36,7 @@ final class NewsletterSubscriptionFormType extends AbstractType
                     'required' => true,
                     'attr' => [
                         'placeholder' => 'newsletter.form.email',
-                        'class' => 'lp-c-light-grey bg-transparent border-2 rounded-0',
+                        'class' => 'lp-c-light-grey border-2 rounded-0',
                     ],
                 ]
             )
@@ -48,7 +48,7 @@ final class NewsletterSubscriptionFormType extends AbstractType
                     'required' => false,
                     'attr' => [
                         'placeholder' => 'newsletter.form.zip',
-                        'class' => 'lp-c-light-grey bg-transparent border-2 rounded-0',
+                        'class' => 'lp-c-light-grey border-2 rounded-0',
                     ],
                 ]
             )
@@ -60,7 +60,7 @@ final class NewsletterSubscriptionFormType extends AbstractType
                     'required' => false,
                     'attr' => [
                         'placeholder' => 'newsletter.form.phone',
-                        'class' => 'lp-c-light-grey bg-transparent border-2 rounded-0',
+                        'class' => 'lp-c-light-grey border-2 rounded-0',
                     ],
                 ]
             )