+++ CHANGELOG.md
@@ -7,3 +7,4 @@ Changelog
  * replace custom Annotation in UrlVimeoConstraint by PHP Attribute
  * apply webp output format into all Liip Imagine filters
  * set AssetMapper usage instead of WebpackEncore
+ * remove bg-transparent from newsletter subscription from due to bad placeholder font color reading